Add files via upload
At the beginning of this project, I had the following question: given a (fixed) pair of nxn matrices A and B what can we learn about the determinants of matrices of the form WAX+YBZ, where W, X, Y, and Z are orthogonal matrices of the same size, particularly as n goes to infinity?

To investigate the question, I decided to have Python randomly generate a large number of matrices of the desired form. In particular, I had Python randomly generate 1000 quintuples of orthogonal W, X, Y, and Z using the orthogonal part of the QR factorization of randomly chosen matrices.

The DeterminantCalculations file has all of the code needed to pull of the linear algebra part of the project. In particular, it provides a function, genData(), that will generate 1000 determinants of large random matrices of the desired form for a fixed A and B.

The GenerateData file runs the genData function five times (each time the function is run, a new A and B are chosen). The file stores all of the data from th